moving bipartite.h to header
